+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
-- TMS-968: Add recurring event fields to manual events
+- TMS-968:
+    - Add recurring event fields to manual events
+    - Add recurring event logic to events component, combined event search & combined event listing
 
 ## [1.2.0] - 2024-02-01